+++ CHANGES.rst
@@ -21,6 +21,10 @@ Bug fixes:
 - Adapt code to follow Plone styleguide.
   [gforcada]
 
+- Fix in catalog: If ZCTextIndex catches a ParseError, try w/o SearchableText if there is one given.
+  Often ``Query contains only common words`` happen (only stop words given) and results are not fetched, then we want all.
+  [jensens]
+
 3.0 (2016-06-12)
 ----------------